% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue.R
\docType{data}
\name{dengue}
\alias{dengue}
\title{Dengue Fever cases in Italy from 2008 to 2021}
\format{
A \link{ts} object with 168 monthly observations.
}
\source{
European Centre for Disease Prevention and Control (ECDC) Surveillance Atlas of Infectious Diseases  \url{https://atlas.ecdc.europa.eu/public/index.aspx}.
}
\usage{
dengue
}
\description{
Monthly cases of Dengue fever registered in Italy from January 2008 to December 2021.
}
\keyword{datasets}
