% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odes.R
\name{print.odes}
\alias{print.odes}
\title{Summarizing Observation-driven Exponential Smoothing}
\usage{
\method{print}{odes}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of \code{odes} class.}

\item{digits}{The number of significant digits to use when printing.}

\item{\\dots}{Not used.}
}
\description{
\code{print} method for \code{odes} objects.
}
\examples{
dengue_data <- data.frame(counts = dengue, time = seq_len(length(dengue)))
fit <- odes(counts ~ sin(2 * pi * time / 12) + cos(2 * pi * time / 12),
data = dengue_data)
fit

}
\references{
{
Karlis, D., Pedeli, X. and Varin, C. (2023). Observation-driven exponential
smoothing for Poisson counts. \emph{Stat}, \strong{12} (1), e642.
}
}
