% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odes.R
\name{summary.odes}
\alias{summary.odes}
\title{Summarizing observation driven exponential smoothing}
\usage{
\method{summary}{odes}(
  object,
  sandwich = TRUE,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{object}{An object of \code{odes} class.}

\item{sandwich}{Logical. Should standard errors be computed with the robust
sandwich HAC variance or not?}

\item{digits}{The number of significant digits to use when printing.}

\item{signif.stars}{Logical. If 'TRUE', \verb{significance stars} are printed for
each coefficient.}

\item{\\dots}{Not used.

#' @references{
Karlis, D., Pedeli, X. and Varin, C. (2023). Observation-driven exponential
smoothing for Poisson counts. \emph{Stat}, \strong{12} (1), e642.
}}
}
\description{
\code{summary} method for \code{odes} objects.
}
\examples{
dengue_data <- data.frame(counts = dengue, time = seq_len(length(dengue)))
fit <- odes(counts ~ sin(2 * pi * time / 12) + cos(2 * pi * time / 12),
data = dengue_data)
## summary with ordinary standard errors
summary(fit, sandwich = FALSE)
## summary with robust sandwich standard errors
summary(fit)

}
