% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odes.R
\name{surveillance.odes}
\alias{surveillance.odes}
\title{Surveillance with Observation-driven Exponential Smoothing}
\usage{
surveillance.odes(
  object,
  new_data,
  level = 0.99,
  plot = TRUE,
  names_as_dates = FALSE,
  x_label = "Time",
  y_label = "Count",
  ...
)
}
\arguments{
\item{object}{An object of \code{odes} class corresponding to a fitted model to
phase I data.}

\item{new_data}{Data frame containing the observations to monitor
corresponding to phase II data.}

\item{level}{Level used in the computation the upper control limit. Default
is \code{0.99}.}

\item{plot}{Logical. Should a graph with the series and the upper control
limit be plotted? Default is \code{TRUE}.}

\item{names_as_dates}{Logical. Are the names of the observations dates?
Default is \code{FALSE}. Only used if \code{plot} is \code{TRUE}.}

\item{x_label}{Character. Label for x-axis. Default is \code{Time}. Only used if
\code{plot} is \code{TRUE}.}

\item{y_label}{Character. Label for y-axis. Default is  \code{Count}. Only used if
\code{plot} is \code{TRUE}.}

\item{\\dots}{Additional optional graphical parameters passed to \code{ggplot}.}
}
\value{
An object of class \code{surveillance.odes}.
}
\description{
Monitors a time series of counts with observation-driven exponential smoothing.
Monitoring is carried out by comparing new observations (phase II data) with
an upper control limit calculated with parameters estimated from historical
data (phase I).
}
\examples{
## Replicates Section 4 of Karlis, Pedeli and Varin (2023)
dengue_data <- data.frame(counts = dengue, time = seq_len(length(dengue)))
dates <- seq(as.Date("2008/1/1"), by = "month", length.out =
nrow(dengue_data))
rownames(dengue_data) <- dates
## phase I
dengue_phaseI <- dengue_data[1:108, ]
fit <- odes(counts ~ sin(2 * pi * time / 12) + cos(2 * pi * time / 12),
data = dengue_phaseI)
summary(fit)
## phase 2
dengue_phaseII <- dengue_data[-(1:108), ]
surv <- surveillance.odes(fit, new_data = dengue_phaseII, names_as_dates =
TRUE, date_labels = "\%Y", breaks = "2 year", x_label = NULL, y_label =
"Dengue cases")
surv$alarms

}
\references{
{
Karlis, D., Pedeli, X. and Varin, C. (2023). Observation-driven exponential
smoothing for Poisson counts. \emph{Stat}, \strong{12} (1), e642.
}
}
